refactor: rename `register` to `registers`
This module contains various registers.
