+++ kernel/src/device/pci/xhci/context.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use super::register::Registers;
+use super::registers::Registers;
 use crate::mem::allocator::page_box::PageBox;
 use alloc::rc::Rc;
 use bitfield::bitfield;
